refactor(server): moving the file location from settings to application
