+++ packages/amplication-client/src/App.tsx
@@ -12,8 +12,8 @@ import PrivateRoute from "./authentication/PrivateRoute";
 import NavigationTabsProvider from "./Layout/NavigationTabsProvider";
 import ThemeProvider from "./Layout/ThemeProvider";
 import { track, dispatch, init as initAnalytics } from "./util/analytics";
-import AuthAppWithGithubCallback from "./Settings/AuthAppWithGithubCallback";
 import RouteWithAnalytics from "./Layout/RouteWithAnalytics";
+import AuthAppWithGithubCallback from "./Application/github/AuthAppWithGithubCallback";
 
 const context = {
   source: "amplication-client",
