+++ packages/amplication-client/src/Application/github/AuthAppWithGithub.scss
@@ -1,4 +1,4 @@
-@import "../style/index.scss";
+@import "../../style/index.scss";
 
 .auth-app-with-github {
   margin-left: 0 !important;
