+++ packages/amplication-client/src/Application/github/AuthAppWithGithubCallback.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { match } from "react-router-dom";
 import { gql, useMutation } from "@apollo/client";
-import { useTracking } from "../util/analytics";
+import { useTracking } from "../../util/analytics";
 
 type Props = {
   match: match<{ application: string }>;
