+++ packages/amplication-client/src/Application/github/SyncWithGithubPage.scss
@@ -1,4 +1,4 @@
-@import "../style/index.scss";
+@import "../../style/index.scss";
 
 $max-width: 700px;
 $icon-font-size: 40px;
